[NIAD-3144] Refactor code to abide to single responsibility principle
